FileAPI/tests/submissions/Breezewish/FileReader-Erros-Test1.html

<!DOCTYPE html>
<html>
<body>
<head>
<meta charset=utf-8>
<title>FileReader Errors Test (1)</title>
<link rel=help href="http://www.w3.org/TR/FileAPI/#convenienceAPI">
<link rel=author title="Breezewish" href="mailto:me@breeswish.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <form name="upload">
        <input type="file" id="fileChooser"><br><input type="button" id="start" value="start">
    </form>

    <div>
        <p>Test steps:</p>
        <ol>
            <li>Download the <a href="support/file_test1.js">file</a>.</li>
            <li>Select the file in the file inputbox.</li>
            <li>Delete the file.</li>
            <li>Click the 'start' button.</li>
        </ol>
    </div>

    <div id="log"></div>

    <script>

        var fileChooser = document.querySelector('#fileChooser');

        setup({explicit_done: true});
        setup({explicit_timeout: true});

        on_event(fileChooser, 'change', function() {

            (function() {

                var t = async_test('FileReader.error should be null if there are no errors when reading');
                t.step(function() {

                    var reader = new FileReader();
                    reader.readAsArrayBuffer(fileChooser.files[0]);

                    reader.onloadend = t.step_func_done(function(event) {
                        if (event.target.readyState == FileReader.DONE) {
                            assert_equals(this.error, null);
                            t.done();
                        }
                    }, reader);

                });

            })();

        });

        on_event(document.querySelector('#start'), 'click', function() {

            (function() {

                var t = async_test('FileReader.error should be NOT_FOUND_ERR if the file is not found when reading');
                t.step(function() {

                    var reader = new FileReader();
                    reader.readAsArrayBuffer(fileChooser.files[0]);

                    reader.onloadend = t.step_func_done(function(event) {
                        if (event.target.readyState == FileReader.DONE) {
                            assert_equals(this.error.code, 1);
                            t.done();
                        }
                    }, reader);

                });

            })();


            done();

        });
        

    </script>
</body>
</html>
